+++ src/api/kotlin/net/ndrei/teslapoweredthingies/api/itemliquefier/IItemLiquefierRegistry.kt
@@ -1,0 +1,8 @@
+package net.ndrei.teslapoweredthingies.api.itemliquefier
+
+import net.minecraft.item.ItemStack
+import net.ndrei.teslapoweredthingies.api.IPoweredRegistry
+
+interface IItemLiquefierRegistry<T: IItemLiquefierRecipe<T>>: IPoweredRegistry<T> {
+    fun getRecipe(input: ItemStack): T?
+}