+++ changelog.md
@@ -11,6 +11,7 @@
 - add a quarry machine
 - add 'strategy' addons for pump and quarry
 - add CT and API support to all non-farming machines (including the ones with Secondary Outputs, might be dependent on the support for that being added to TCL)
+- add power cost to item liquefier recipes
 
 #1.0.11
 - fixed crop cloner seed duplication bug